server/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
motor==3.3.1
transformers>=4.35.0
torch>=2.0.0
Pillow>=10.0.0
deep-translator>=1.11.4
gtts>=2.4.0
python-multipart>=0.0.6
